fix for mpeg2_aac support
ADTS header for aac was not created correctly when using
mpeg2_aac_low profile. The problem was caused by forcing profile
to be acc_low before it reached the part where adts header is
created. By the time when adts header is created it is impossible
to know whether the profile was mpeg2_aac_low or simply aac_low.