Add IDR frame test logic in segment muxer.
NVENC is failing to provide IDR frame information with
AV_PKT_FLAG_KEY set and causing the segment muxer to fail to
segment the video. The frame packet is examined by the segment
muxer by parsing the nal unit to see if the current frame is an
IDR frame and set the flag manually.

This should be a temporary fix.I personally think that NVIDIA is
at fault and should fix this issue. Otherwise, the encoded packet
should be examined in nvenc.c right after the encoded packet is
received from the ecndoer.